Refactor API layer using Facade pattern (#67)
